Use web_url instead of path in links_to_sentence
The Content API returns tags with `web_url` properties, so if we use
this we don't need to wrap the organisation tags in presenters (nor will
we need to wrap future tags in presenters, such as browse sections and
specialist sectors).

Also update the method signature for `links_to_sentence` to not set a
(misleading) default value for its argument.