404 if the Repo can't find a document
With the introduction of a Repo for fetching documents, the Repo would
throw an error if params[:section_id] didn't match a placeholder file
instead of allowing the controller to respond with a 404. This commit
checks for the existence of a file first, which if false allows the
controller to respond with a 404.