Limit width of Manual and Document summaries
This commit limits the width of Manual and Document summaries in order
to bring it in line with the rest of the body content in Documents.

This also fixes an issue with the bug fix for line width on content in
the body. It does this by using the CSS not selector to go through all
children except h2s, so it will never skip the first element in
.govspeak, which might be a h2.