Attempt to serve employment-income-manual
If it exists in content-store, then show it as normal. If it doesn't, show the
holding page as before.

This is useful because the employment-income-manual is the one that the HMRC
Manuals API team use, and also that HMRC/their developer iO1 use too. This
should also remove a small burden when it comes time for this manual to be
published for real.