Use print.css when printing
After the print stylesheet was added in 0111c5f it was being included
with the stylesheet_link_tag helper as print.css. The helper then
thought this was at the route and was including it as /print.css at the
root of the domain. This commit removes the .css which allows the helper
to function correctly.