Unify styles between index and show pages
Show and Index have a lot in common above the main body content (they both have a Summary and an optional Body which should be two-thirds width at max screen size)

This commit removes some divs that weren't necessary, and shares styles between these parts of the page so there is less duplication of css.

This commit renames 'manual-section' to 'manual-body' which seems like a more understandable name.

This commit also refactors the index page into a list of sections rather than a lot of stacked headers (which it was before because it had to accommodate different depths of content, so some would be links to other pages and some would have copy underneath them which folded out when you clicked the header - yuck.)