Fix up styles for updates page
Given the flexible and highly changable length of the change discriptions and the affected page names, in terms of layouts this is about as nice as we're going to get while still accommodating all text lengths. This commit adds some styling for updates lists and removes some (now unused) styles for the old updates presentation.