Fix Firefox bug
In Firefox the page length is calculated by the content even if that content is off the page (e.g. at a negative margin). This is different to the way all IE/Chrome do it and was only just spotted. Add 1px overflow hidden to the hidden sections so FF calculates the page height as we want and we don't get a really huge blank footer.