Fix view logic for HMRC leaf nodes
As HMRC leaf nodes are now on the document.body, make sure they aren't accidentally made collapsible.
I'm not sure if it's possible for the HMRC API to give us a document with a `document.body` AND `document.section_groups` but it seems sensible to show them both if they're both there.

Non HMRC manuals should just have a `document.body` on their show action, but they don't need a don't need to be wrapped in a `.body-content-wrapper` class because their headings still need to be full width (as visually they are styled differently to hrmc's manuals)

Further work on this file (which I'm leaving here as a note for myself later) would be to remove:
`
<div class='title-controls-wrap'></div>
<div class='collapsible-subsections'>
`
and leave add the javascript to add in these elements as they are not needed unless the javascript is running.