Fix the alignment of the titles
This is one of those annoying CSS problems that it is impossible to fix fully without sacrificing semantics crucial for accessibility.

This commit gives the `.subsection-id` field a minimum width that brings the title inline with the metadata in the head, unless this code is longer than 135px, in which case it will push further out. There are some manuals with longer subsection ids than 135px at font-size: 19, e.g. http://www.hmrc.gov.uk/manuals/sgsaroa/index.htm, however these are less common.

An alternative solution was investigated whereby the table layout algorithms were used (i.e. `display:table-row` was applied to the `li`) however this only works if the elements that have `display:table-cell` have only one parent, to which you apply `display:table-row`. Since in this case the id and title divs have a an `<a>` and an `<li>` parenting them, this approach doesn't work. A possible solution of adding links around the individual id and title was looked at but this does quite odd things with keyboard focus unless you write some Javascript to correct focus behaviour.

In short: this solution is perfect for most titles and good enough for the rest without sacrificing accessibility for the sake of pixel perfection.