Add article tag and labelled by
The index pages are the only pages of the manual whose h1 is the h1 in the header. Add the article tag and use labelledby to point to the h1 of the page in the header. This markup is different to the show/updates pages were the header is less important than on the index page