Add article, labelledby + h1 to show + updates page
A fix for: https://github.com/alphagov/manuals-frontend/issues/104

Currently, all of the pages in a manual have the same h1- the title of the manual. This isn't the most sensible markup as it means the h1 is basically useless and the h2 is the most useful heading.

This fix uses HTML5 style headings whereby h1s are scoped to their container element. So the h1 in the header is the highest order heading for the header scope, and by adding in an article tag, we can make the page title an h1 but have it as the highest order heading for the article.