Build open/close html with JS
Having removed the HTML from the show and updates templates, we need to build it with the javascript.

This commit makes the javascript add in `<div class='js-title-controls-wrap'>` which is the parent node to the one that we were already adding here.

This code also handles wrapping the sections title if it's there (its currently only used on the updates page)