Improve README
This readme didn't contain enough information for someone to get started with the app. This is a start though I think it could still be improved. We're about to get a new lot of devs committing to this repo, I'd like to improve this readme as we go, but this is a start.