Add document repository and sibling decorator
Ideally previous and next links data would be provided by the Content Store and
it would be calculate from the Sibling Decorator.
If and when the Content Store will be able to do so, we will need to only delete
the Sibling Decorator in order to remove the sibling calculation logic.

Added the document repository to allow to query the Content Store only from within
it instead of doing so from many different parts of the code.

Manuals Controller:
- exposed document and manual paths
- removed the method that fetched content straight from the content store
- added a document repository method

Document Repository:
Mediates between the Sibling Decorator, the Manuals Controller and the Content Store
- fetching parent and the current document from the Content Store
- feeding the document and parent to the Sibling Decorator

Sibling Decorator:
- takes care of the logic needed to find previous and next siblings of the document
- decorates the document with the previous and next siblings methods