Add class link to make print stylesheet targeting more specific
To help future proof the hiding of the content of .secondary div
for print I've added a class to the child element so that it can be
singled out. Thus reducing future work if/when other elements get
added to .secondary that need to be included in the print version
in future.