Remove employment income manual code
This manual was used as a placeholder to demonstrate Manuals Frontend
before we had any content. The placeholder is no longer being used so
this commit removes the `before_filter`, the tests, the views and
changes the rake task which publishes the route for one which
publishes a gone route.