Disable caching for content-store
We shouldn't cache content-store items in the controller requests, as
this results in an inconsistency between what is published in the API
and what is available to clients hitting the controllers directly.

Further to this, Rails's built-in caching is opaque to external systems,
and requires a hard-restart before being cleared. This gives us no
options in the future to easily propagate content-store changes to
downstream clients.