Enable rich govspeak
With the associated change in the Govspeak component, this commit
updates the call to it to set `rich_govspeak` to be `true`. This then
allows Manuals to have bold text within the body. This is required for
the Highway Code.

See https://github.com/alphagov/static/pull/657 for more information.

[Ticket](https://trello.com/c/eYyOkNRR/131-highway-code-manual-has-styled-strong-tags).