Don't return a gone content item
When viewing `/updates` page for a withdrawn content item,
`ManualsController` passes the `gone` content item to the
`ManualPresenter` which fails when trying to call methods which the
`gone` content item does not have, but the `manual` does. This commit
changes the controller to not return the content item if its `format` is
`gone`, allowing `/updates` to 410 as expected.