Upgrade Slimmer, explicitly set layout
This is a no-op change to app behaviour. The same Slimmer layout is used as be
before, but set in a simple, standardised way with `slimmer_template`, which is
easier to grep for than the other methods.

Slimmer 9.0.0 switched to `core_layout` [1] as the default layout, which is the
modern/recommend layout. Apps not yet using `core_layout` should be explicitly
marked as such, so we can easily identify them as needing updating.

Switching to core looks like it would be straightforward, removing the use of
`.inner-block` to set the base grid, but i'd want to do that in a separate PR.

[1] https://github.com/alphagov/slimmer/pull/136