Refactor to remove creation of new collapsible
Previously, when finding a section to open for a given anchor, a new Collapsible was created. This led to duplication of DOM elements and buggy behaviour. This commit fixes that by searching for the Collapsible to be opened among the existing Collapsibles.

This commit also refactors parts of the code in the constructor method that was checking for a hash and opening the corresponding Collapsible for it.