Update manual presenter to prefer links for organisation data
We're currently updating apps to support a new tagging architecture.
Part of this work includes changes to hmrc-manuals-api such that it sets
the organisation data in the links hash rather than the details hash.

This change to manuals-frontend ensures organisation data in the links
field of the manual payload is used in preference to data in details.
This will eventually be updated to drop usage of details altogether.