Add meta descriptions
The meta description tag is used for the index page of a manual but none of the sub-pages. This commit adds meta description tags to all pages, the contents of which are set to the description associated with the manual. This adds descriptions to external search engine results when the pages are re-indexed.