Handle a manual's public_updated_at being null
We've recently changed manuals-publisher so that it no longer sends dates
to the publishing-api when writing drafts or publishing.  This means that
the publishing-api handles all the updated_at, first_published_at, and
public_updated_at values that appear in the content store.  One side-effect
of this is that drafts of a manual have a null value for the
public_updated_at in the content store.  Note that we expect this for the
first draft of a new manual, but we're seeing it for other drafts too.

To avoid breaking manuals-frontend in this situation we add some guard
clauses to the presenter and helpers.