Move the presenters into the models director
They're not presenters per se, where a presenter is usually something that
is used to present a model to the view, these are more like models whose
backend represntation happens to be a JSON api, instead of a traditional
database.

I want to move these because I need to introduce "true" presenters that
have a view_context and can call view helper styls methods.