(Re)Introduce ManualPresenter
This version of the presenter takes a view context and this allows us to
wrap up helper methods in the context of an object.  We use it to turn
updated_at into a version of the date marked up as html, and to generate
the html details that go into the arguments for the metadata govuk
component.