Refactor manuals controller and views
The intention of this commit is to be more explicit about the
requirements of each of the manuals' views. From the perspective of
someone new adding more functionality to this application, I think it
helps to see exactly the dependencies needed form the controller.

I thought it would also be useful to rename some of the variables used
in the views as well in order to make it easier to understand what
objects/data they hold.