Service workers prototype for firebreak
This adds a 'Save for offline use' button to the Style Guide
manual that posts a message to a service worker that precaches
the needed resources (content, styles, scripts) for a
'good enough' experience when offline.

This prototype is very rough and ready, but it gives a basic
idea of how it could work.