Return a redirect when the content-store returns one
When a section is withdrawn from a manual it gets saved as a
redirect in content-store.

There exists a race condition where the router retrieves the content
as being a manual, then before the manuals-frontend requests the document,
publishing-api updates the type to redirect.  This race condition happens
relatively often in the E2E tests.

This change will redirect the request properly instead of raising an
exception inside the Manual model.