Don't include govuk_toolkit in application.js
When the application is deployed we get the govuk_toolkit js files from
static, if we also include them in our own JS bundle we:

1. make users effectively download the same code twice
2. run the risk of delivering two different versions of the code and
   causing errors in how things interact
3. run the risk of delivering two versions of the code that interact
   strangely because they end up redefininig things and cause things
   like type-checks to fail

Of course, if we rely on static to deliver this JS we won't have it in
test environments, so we'd need to do something else to make parts of that
JS bundle available during test.  However the only bit of govuk_toolkit we
actually need is primary-links, which isn't delivered by the layout we use
from static (header-footer-only), so instead of `require govuk_toolkit` we
can instead use `require govuk/primary-links` to pull in only that part of
the gem.  We've included a comment with the require to make it clear that
we should be circumspect in how much of govuk_toolkit we require.

It's worth noting that 2. and 3. are not hypothetical. When we added PII
stripping functionality to govuk_frontend_toolkit and delivered it via
static we saw both of these problems. 1st: the version of the JS included
in the frontend apps was out of sync with the version delivered by static
so it didn't understand the PIISafe object at all and would raise errors
when trying to call that function, resulting in no analytics being sent.
2nd: after updating the govuk_frontend_toolkit version in the apps so it
was in sync with that delivered by static, we saw custom dimension values
of [object Object] being sent to analytics because the two versions on
the page disagreed about the type of Analytics.PIISafe instances and
didn't extract the value correctly.

Removing the govuk_toolkit reqire from our application.js and only
requiring the parts we need that aren't delivered avoids this entirely.