Add machine readable metadata to manuals
This adds the standard `Article` schema to the manual page (not sections
as yet).

We may look at adding the `HasPart` and `IsPartOf` schemas to manuals and
sections, but we'll plan this separately.

Adding a `SearchAction` to manuals pages is coming soon, and will happen
automatically when the publishing components gem is updated with the changes.

https://trello.com/c/4EAytHVV/747-plan-indexing-of-finder-pages