Switch asset host env var for change of asset host
This allows asset host to be set by ASSET_HOST env var rather than
GOVUK_ASSET_HOST. This is in preparation for the assets of this app
changing hostname from assets.publishing.service.gov.uk to
www.gov.uk.

As part of this change the path to assets is changed to
/assets/manuals-frontend

The expectation is that this env var will only be set in dev
environments when proxying isn't set up for router.