Fix date ordering on manuals updates page
This commit fixes date ordering on the manuals updates page. Previously, HMRC manuals displayed updates from earliest to most recent, whereas other manuals displayed updates from most recent to the earliest.

The reason for this is because the change notes for HMRC manuals, as recorded in the content item, ranged from most recent to earliest, whereas for other manuals was from earliest to most recent. A `reverse` operation was then applied to these change notes in the code, which meant that HMRC manuals were then displayed in the wrong order.

To resolve this, instead sort by date descending to ensure ordering is consistent for different manuals.