Return 406 Not Acceptable when 'Accepts' header requests non-HTML
We've had a number of Sentry issues to do with requesting JSON or
JS versions of Manuals Frontend pages:
- https://sentry.io/organizations/govuk/issues/2152998596/?project=202235&query=is%3Aunresolved
- https://sentry.io/organizations/govuk/issues/2161602206/?project=202235&query=is%3Aunresolved

The correct thing to do is to respond with a [406](https://httpstatuses.com/406),
and not to log inactionable Sentry issues.

I've taken inspiration from Content Publisher for the request
testing, and the underlying fix is the same approach as in
government-frontend:
https://github.com/alphagov/government-frontend/blob/ce26ae94d66d9ce92ac37028854211741ef44b8b/app/controllers/content_items_controller.rb#L6-L7

We will also need to do some work to ensure that 406 responses
are cached in Fastly.

Trello: https://trello.com/c/XNQhMGRt/2391-3-make-manuals-frontend-respond-to-bad-requests-with-406-status