ManualSections that are"visually_expanded" render without accordions
Manual sections that are published as `"visually_expanded": true`
will render as flat text  without the "accordion" govuk_publishing_component.
https://components.publishing.service.gov.uk/component-guide/accordion

The default publishing option is `"visually_expanded": false`. If the
visually_expanded attribute is not set the section will also continue
to display as an accordion.

Also removes an unused `collapsible?` method from the Section model.