Prevent title from appearing twice in manual_section
A change to add the breadcrumb to the document heading for
manual_section view has caused the title to sometimes appear twice.

The document_heading is intended to present the `section_id - title`.
The intention is that the `breadcrumb` method should provide the `section_id` - however
if this is not present it nstead defaults to `title`.

https://github.com/alphagov/manuals-frontend/blob/63c17fff271e85fcf9f6196963e0d4135214c29f/app/models/document.rb#L26

This change prevents the title being added to the document_heading if it
is already present i.e: the `breadcrumb` method was unable to return a `section_id`.