Handle 'Gone' documents
Previously the code would raise an error if `document_from_repository`
returned a document that had a status of 'Gone'. This was raised as
[HTTPGone][1] from gds-api-adapters as a result of receiving a 410 from
content-store.

Now 'Gone' pages and their corresponding exception are rescued and
handled by returning a 410 status code which in turn renders a [generic
410 error page][2].

I've also refactored the error handling logic as part of this work
which:

* moves all this behaviour to the manuals controller
* follows the pattern in [government-frontend][3]

[1]: https://github.com/alphagov/gds-api-adapters/blob/master/lib/gds_api/exceptions.rb#L90-L91
[2]: https://github.com/alphagov/static/blob/master/app/views/root/410.html.erb
[3]: https://github.com/alphagov/government-frontend/blob/e5c0c97dc614b9513b82e106074e9293826498a4/app/controllers/content_items_controller.rb#L2-L10